NDP [MAINTENANCE]: change behavior of metadata field in DMA Calypte header
The "metadata" field in the DMA header used on Calypte controller was
previously intended as a user metadata field but this has two
implications. The field was of fixed size of 24 bits which does not
allow to transport longer metadata per frame and secondly, the API does
not provide any way to even write data to this field.

Therefore, this field has been renamed to "metadata_len" and currently
stores the size of metadata in bytes. Metadata are now prepended before
the frame and transported with it. The "header" and "data" pointers
inside each member of the "packets" now contains correct addresses
inside the packet to write "metadata" or "data" into.